Move logger to header
